Separate CMR class in cmrjs.
Fix circular imports.
Rename e => error.
